refactor: lazy ProfilePage
