+++ wondrous-app/pages/organization/[username]/project.tsx
@@ -1,50 +1,11 @@
-import { useLazyQuery, useQuery } from '@apollo/client';
 import { withAuth } from 'components/Auth/withAuth';
-import EntitySidebar from 'components/Common/SidebarEntity';
-import OrgProject from 'components/organization/project';
-import { GET_ORG_FROM_USERNAME, GET_USER_PERMISSION_CONTEXT } from 'graphql/queries';
-import { useRouter } from 'next/router';
-import { useEffect, useMemo } from 'react';
-import { OrgBoardContext } from 'utils/contexts';
+import BoardSkeleton from 'components/Dashboard/boards/BoardSkeleton';
+import lazy from 'utils/enhancements/lazy';
 
-const useGetOrgFromUsername = (username) => {
-  const [getOrgFromUsername, { data }] = useLazyQuery(GET_ORG_FROM_USERNAME);
-  useEffect(() => {
-    if (!data && username) {
-      getOrgFromUsername({
-        variables: {
-          username,
-        },
-      });
-    }
-  }, [username, data, getOrgFromUsername]);
-  return data?.getOrgFromUsername;
-};
+const OrgProject = lazy(() => import('components/organization/project'), BoardSkeleton);
 
-function ActivitiesPage() {
-  const router = useRouter();
-  const { username } = router.query;
-  const { data: userPermissionsContext } = useQuery(GET_USER_PERMISSION_CONTEXT, {
-    fetchPolicy: 'cache-and-network',
-  });
-  const org = useGetOrgFromUsername(username);
-  const contextValue = useMemo(
-    () => ({
-      userPermissionsContext: userPermissionsContext?.getUserPermissionContext
-        ? JSON.parse(userPermissionsContext?.getUserPermissionContext)
-        : null,
-      orgId: org?.id,
-      orgData: org,
-    }),
-    [org, userPermissionsContext?.getUserPermissionContext]
-  );
-  return (
-    <OrgBoardContext.Provider value={contextValue}>
-      <EntitySidebar>
-        <OrgProject orgData={org} />
-      </EntitySidebar>
-    </OrgBoardContext.Provider>
-  );
+function ProjectPage(props) {
+  return <OrgProject {...props} />;
 }
 
-export default withAuth(ActivitiesPage);
+export default withAuth(ProjectPage);
